Organize paths of core code files
