+++ src/api/grpc/server/spec.go
@@ -7,8 +7,8 @@ import (
 	"github.com/gogo/protobuf/types"
 
 	pb "github.com/cloud-barista/cb-tumblebug/src/api/grpc/protobuf"
-	"github.com/cloud-barista/cb-tumblebug/src/common"
-	"github.com/cloud-barista/cb-tumblebug/src/mcir"
+	"github.com/cloud-barista/cb-tumblebug/src/core/common"
+	"github.com/cloud-barista/cb-tumblebug/src/core/mcir"
 )
 
 func (s *server) RegisterSpecWithCspSpecName(ctx context.Context, req *pb.RegisterSpecWithCspSpecNameWrapper) (*pb.TbSpecInfo, error) {
