+++ src/cli/cmd/root.go
@@ -27,8 +27,8 @@ import (
 
 	//conf "github.com/cloud-barista/cb-tumblebug/src/cli/config"
 
-	"github.com/cloud-barista/cb-tumblebug/src/common"
-	"github.com/cloud-barista/cb-tumblebug/src/mcir"
+	"github.com/cloud-barista/cb-tumblebug/src/core/common"
+	"github.com/cloud-barista/cb-tumblebug/src/core/mcir"
 )
 
 const ActionTerminate string = "Terminate"
