+++ src/core/mcir/obsolete/subnet.go
@@ -7,7 +7,7 @@ import (
 
 	//"strings"
 
-	"github.com/cloud-barista/cb-tumblebug/src/common"
+	"github.com/cloud-barista/cb-tumblebug/src/core/common"
 	"github.com/labstack/echo/v4"
 )
