+++ src/core/mcis/control.go
@@ -25,8 +25,8 @@ import (
 
 	"sync"
 
-	"github.com/cloud-barista/cb-tumblebug/src/common"
-	"github.com/cloud-barista/cb-tumblebug/src/mcir"
+	"github.com/cloud-barista/cb-tumblebug/src/core/common"
+	"github.com/cloud-barista/cb-tumblebug/src/core/mcir"
 )
 
 const ActionCreate string = "Create"
