+++ src/core/mcis/monitor.go
@@ -19,7 +19,7 @@ import (
 
 	"sync"
 
-	"github.com/cloud-barista/cb-tumblebug/src/common"
+	"github.com/cloud-barista/cb-tumblebug/src/core/common"
 
 )
